+++ horovod/ray/runner.py
@@ -1,18 +1,16 @@
 import ray
-from ray import services
 
+import warnings
 from collections import defaultdict
 from dataclasses import dataclass
 import os
-import socket
 from typing import Dict, Callable, Any, Optional, List
 import logging
 
 from horovod.runner.common.util import secret, timeout, hosts
 from horovod.runner.http.http_server import RendezvousServer
-from horovod.ray import ray_logger
-from horovod.ray.utils import detect_nics, nics_to_env_var
-
+from horovod.ray.utils import detect_nics, nics_to_env_var, map_blocking
+from horovod.ray.strategy import ColocatedStrategy, PackStrategy
 logger = logging.getLogger(__name__)
 
 
@@ -38,67 +36,6 @@ class MiniSettings:
             "check connectivity between servers. You "
             "may need to increase the --start-timeout "
             "parameter if you have too many servers.")
-
-
-def map_blocking(fn, collection):
-    return ray.get([fn(w) for w in collection])
-
-
-class BaseHorovodWorker:
-    executable = None
-
-    def __init__(self, world_rank=0, world_size=1):
-        os.environ["HOROVOD_HOSTNAME"] = self.hostname()
-        os.environ["HOROVOD_RANK"] = str(world_rank)
-        os.environ["HOROVOD_SIZE"] = str(world_size)
-
-    def hostname(self) -> str:
-        # TODO: This is probably not the right way to retrieve
-        # the intended hostname.
-        return socket.gethostname()
-
-    def get_gpu_ids(self) -> List[int]:
-        """Return list of CUDA device IDs available to this worker."""
-        return ray.get_gpu_ids()
-
-    def update_env_vars(self, env_vars: Dict[str, str]):
-        """Update the env vars in the actor process."""
-        sanitized = {k: str(v) for k, v in env_vars.items()}
-        os.environ.update(sanitized)
-
-    def env_vars(self):
-        """Check the env vars in the actor process."""
-        return dict(os.environ)
-
-    def start_executable(self,
-                         executable_cls: type = None,
-                         executable_args: list = None,
-                         executable_kwargs: dict = None):
-        """Instantiates the executable class with provided args.
-
-        If none, self.executable = None.
-
-        Args:
-            executable_cls (type): Class of object to be created on all
-                workers.
-            executable_args (list): Initialization arguments for the
-                executable_cls.
-            executable_kwargs (dict): Initialization arguments for the
-                executable_cls.
-        """
-        executable_args = executable_args or []
-        executable_kwargs = executable_kwargs or {}
-        if executable_cls:
-            self.executable = executable_cls(*executable_args,
-                                             **executable_kwargs)
-
-    def execute(self, func):
-        """Executes an arbitrary function on self."""
-        return func(self.executable)
-
-    def set_queue(self, queue):
-        """Sets the queue for multi-node logging."""
-        ray_logger.configure(queue=queue)
 
 
 class Coordinator:
@@ -135,12 +72,20 @@ class Coordinator:
     def finalize_registration(self) -> dict:
         """Return a dictionary for all ranks."""
         rank_to_info = {}
+
+        cross_sizes = defaultdict(int)
+        cross_ranks = {}
+        for rank_list in self.hostnames_by_rank.values():
+            for local_rank, world_rank in enumerate(rank_list):
+                cross_ranks[world_rank] = cross_sizes[local_rank]
+                cross_sizes[local_rank] += 1
+
         for node_world_rank, (hostname, ranks) in enumerate(
                 self.hostnames_by_rank.items()):
             for local_rank, world_rank in enumerate(ranks):
                 rank_to_info[world_rank] = dict(
-                    HOROVOD_CROSS_RANK=node_world_rank,
-                    HOROVOD_CROSS_SIZE=len(self.hostnames_by_rank),
+                    HOROVOD_CROSS_RANK=cross_ranks[world_rank],
+                    HOROVOD_CROSS_SIZE=cross_sizes[local_rank],
                     HOROVOD_LOCAL_RANK=local_rank,
                     HOROVOD_LOCAL_SIZE=len(ranks))
         return rank_to_info
@@ -166,7 +111,7 @@ class Coordinator:
         self.rendezvous.init(host_alloc_plan)
 
         return {
-            "HOROVOD_GLOO_RENDEZVOUS_ADDR": services.get_node_ip_address(),
+            "HOROVOD_GLOO_RENDEZVOUS_ADDR": ray.util.get_node_ip_address(),
             "HOROVOD_GLOO_RENDEZVOUS_PORT": str(self.global_rendezv_port),
             "HOROVOD_CONTROLLER": "gloo",
             "HOROVOD_CPU_OPERATIONS": "gloo",
@@ -180,14 +125,21 @@ class RayExecutor:
         settings (horovod.Settings): Configuration for job setup. You can
             use a standard Horovod Settings object or create one directly
             from RayExecutor.create_settings.
-        num_hosts (int): Number of machines to execute the job on.
-        num_slots (int): Number of workers to be placed on each machine.
-        cpus_per_slot (int): Number of CPU resources to allocate to
+        num_workers (int): Number of workers to use for training.
+        cpus_per_worker (int): Number of CPU resources to allocate to
             each worker.
         use_gpu (bool): Whether to use GPU for allocation. TODO: this
             can be removed.
-        gpus_per_slot (int): Number of GPU resources to allocate to
+        gpus_per_worker (int): Number of GPU resources to allocate to
             each worker.
+        num_hosts (int): Alternative API to ``num_workers``. Number of
+            machines to execute the job on. Used to enforce equal number of
+            workers on each machine.
+        num_workers_per_host (int): Alternative API to
+            ``num_workers``. Number of workers to be placed on each machine.
+            Used to enforce equal number of workers on each machine. Only
+            used in conjunction with `num_hosts`.
+
     """
 
     @classmethod
@@ -219,96 +171,69 @@ class RayExecutor:
             timeout_s=timeout_s,
             placement_group_timeout_s=placement_group_timeout_s)
 
-    def __init__(self,
-                 settings,
-                 num_hosts: int = 1,
-                 num_slots: int = 1,
-                 cpus_per_slot: int = 1,
-                 use_gpu: bool = False,
-                 gpus_per_slot: Optional[int] = None):
+    def __init__(
+            self,
+            settings,
+            num_workers: Optional[int] = None,
+            num_hosts: Optional[int] = None,
+            num_workers_per_host: int = 1,
+            cpus_per_worker: int = 1,
+            use_gpu: bool = False,
+            gpus_per_worker: Optional[int] = None,
+            # Deprecated Args.
+            num_slots: Optional[int] = None,
+            cpus_per_slot: Optional[int] = None,
+            gpus_per_slot: Optional[int] = None):
 
-        if gpus_per_slot and not use_gpu:
-            raise ValueError("gpus_per_slot is set, but use_gpu is False. "
-                             "use_gpu must be True if gpus_per_slot is set. ")
-        if use_gpu and isinstance(gpus_per_slot, int) and gpus_per_slot < 1:
+        if num_slots:
+            warnings.warn(
+                "`num_slots` is now deprecated. Please use the `num_workers` "
+                "API, or to enforce an equal number of workers on each node, "
+                "set `num_hosts` and `num_workers_per_host`. "
+                "This will raise an error in a later release of Horovod. "
+                "Setting num_workers_per_host = num_slots.",
+                category=DeprecationWarning,
+                stacklevel=2)
+            num_workers_per_host = num_slots
+
+        if cpus_per_slot or gpus_per_slot:
+            warnings.warn(
+                "`cpus_per_slot` and `gpus_per_slot` have been deprecated. "
+                "Use `cpus_per_worker` and `gpus_per_worker` instead. "
+                "This will raise an error in a later release of Horovod. "
+                "Setting cpus/gpus_per_slot = cpus/gpus_per_worker.",
+                category=DeprecationWarning,
+                stacklevel=2)
+            cpus_per_worker = cpus_per_slot
+            gpus_per_worker = gpus_per_slot
+
+        if num_workers is None and num_hosts is None:
+            raise ValueError("Either `num_workers` or `num_hosts` must be "
+                             "set.")
+
+        if num_workers and num_hosts:
+            raise ValueError("Both `num_workers` and `num_hosts` cannot be "
+                             "set.")
+
+        if gpus_per_worker and not use_gpu:
+            raise ValueError("gpus_per_worker is set, but use_gpu is False. "
+                             "use_gpu must be True if gpus_per_worker is "
+                             "set. ")
+        if use_gpu and isinstance(gpus_per_worker,
+                                  int) and gpus_per_worker < 1:
             raise ValueError(
-                f"gpus_per_slot must be >= 1: Got {gpus_per_slot}.")
+                f"gpus_per_worker must be >= 1: Got {gpus_per_worker}.")
 
         self.settings = settings
+        self.num_workers = num_workers
         self.num_hosts = num_hosts
-        self.num_slots = num_slots
-        self.cpus_per_slot = cpus_per_slot
+        self.num_workers_per_host = num_workers_per_host
+        self.cpus_per_worker = cpus_per_worker
         self.use_gpu = use_gpu
-        self.gpus_per_slot = gpus_per_slot or 1
+        self.gpus_per_worker = gpus_per_worker or 1
 
         self.workers = []
-        self.placement_group = None
-
-    @property
-    def num_workers(self):
-        return self.num_hosts * self.num_slots
-
-    def _create_workers(self, host_resources):
-        bundles = [host_resources.copy() for _ in range(self.num_hosts)]
-        pg = ray.util.placement_group(bundles, strategy="STRICT_SPREAD")
-        self.placement_group = pg
-        logger.debug("Waiting for placement group to start.")
-        ready, _ = ray.wait(
-            [pg.ready()], timeout=self.settings.placement_group_timeout_s)
-        if ready:
-            logger.debug("Placement group has started.")
-        else:
-            raise TimeoutError(
-                "Placement group creation timed out. Make sure "
-                "your cluster either has enough resources or use "
-                "an autoscaling cluster. Current resources "
-                "available: {}, resources requested by the "
-                "placement group: {}".format(ray.available_resources(),
-                                             pg.bundle_specs))
-
-        # Placement group has started. Now create the workers.
-        self.workers = []
-        # Keep ref of one worker per node for NIC detection.
-        node_workers = []
-        # STRICT_SPREAD guarantees each bundle is on a different node.
-        # Create num_slots workers per bundle, i.e. per machine.
-        for bundle_index in range(len(bundles)):
-            gpu_id_futures = []
-            curr_node_workers = []
-            remote_cls = ray.remote(BaseHorovodWorker)
-            for i in range(self.num_slots):
-                remote_cls_with_options = remote_cls.options(
-                    num_cpus=self.cpus_per_slot,
-                    num_gpus=self.gpus_per_slot * int(self.use_gpu),
-                    placement_group=pg,
-                    placement_group_bundle_index=bundle_index)
-                worker = remote_cls_with_options.remote(
-                    world_rank=self.num_slots * bundle_index + i,
-                    world_size=self.num_workers)
-                if self.use_gpu:
-                    gpu_id_futures.append(worker.get_gpu_ids.remote())
-                self.workers.append(worker)
-                curr_node_workers.append(worker)
-            if len(gpu_id_futures) > 0:
-                # By setting CUDA VISIBLE DEVICES to ALL GPUs,
-                # CUDA will be able to detect adjacent devices and use IPC
-                # allowing for better performance.
-                gpu_ids = sum(ray.get(gpu_id_futures), [])
-                # Make sure that each worker on the node has unique device.
-                assert len(gpu_ids) == len(
-                    set(gpu_ids)) == self.num_slots, gpu_ids
-                all_ids = ",".join([str(gpu_id) for gpu_id in gpu_ids])
-                futures = []
-                for worker in curr_node_workers:
-                    futures.append(
-                        worker.update_env_vars.remote({
-                            "CUDA_VISIBLE_DEVICES":
-                            all_ids
-                        }))
-                ray.get(futures)
-            node_workers.append(curr_node_workers[0])
-
-        return self.workers, node_workers
+        self.strategy = None
 
     def _start_executables(self, executable_cls, executable_args,
                            executable_kwargs):
@@ -317,6 +242,24 @@ class RayExecutor:
                 executable_cls, executable_args, executable_kwargs)
 
         map_blocking(_start_exec, self.workers)
+
+    def _create_strategy(self):
+        assert self.num_workers is None or self.num_hosts is None
+        if self.num_workers:
+            return PackStrategy(
+                settings=self.settings,
+                num_workers=self.num_workers,
+                use_gpu=self.use_gpu,
+                cpus_per_worker=self.cpus_per_worker,
+                gpus_per_worker=self.gpus_per_worker)
+        else:
+            return ColocatedStrategy(
+                settings=self.settings,
+                num_hosts=self.num_hosts,
+                num_workers_per_host=self.num_workers_per_host,
+                use_gpu=self.use_gpu,
+                cpus_per_worker=self.cpus_per_worker,
+                gpus_per_worker=self.gpus_per_worker)
 
     def start(self,
               executable_cls: type = None,
@@ -344,14 +287,10 @@ class RayExecutor:
         """
         extra_env_vars = extra_env_vars or {}
 
-        def resources_per_host():
-            num_cpus = self.cpus_per_slot * self.num_slots
-            num_gpus = self.gpus_per_slot * self.num_slots * int(self.use_gpu)
-            return dict(CPU=num_cpus, GPU=num_gpus)
-
+        self.strategy = self._create_strategy()
         self.coordinator = Coordinator(self.settings)
         executable_args = executable_args or []
-        self.workers, node_workers = self._create_workers(resources_per_host())
+        self.workers, node_workers = self.strategy.create_workers()
         # Get all the hostnames of all workers
         hostnames = map_blocking(lambda w: w.hostname.remote(), self.workers)
         # Register each hostname to the coordinator. assumes the hostname
@@ -448,8 +387,5 @@ class RayExecutor:
         for worker in self.workers:
             del worker
 
-        if self.placement_group:
-            ray.util.remove_placement_group(self.placement_group)
-
-        self.workers = []
-        self.placement_group = None
+        if self.strategy:
+            self.strategy.shutdown()