+++ test/single/test_ray.py
@@ -8,12 +8,11 @@ import sys
 import socket
 import pytest
 import ray
-from ray import services
 import torch
 
 from horovod.common.util import gloo_built
-from horovod.ray.runner import (BaseHorovodWorker, Coordinator,
-                                MiniSettings, RayExecutor)
+from horovod.ray.runner import (Coordinator, MiniSettings, RayExecutor)
+from horovod.ray.worker import BaseHorovodWorker
 
 sys.path.append(os.path.dirname(__file__))
 
@@ -87,23 +86,71 @@ def test_coordinator_registration():
             for info in rank_to_info.values()} == {0, 1, 2, 3}
 
 
-def test_infeasible_placement(ray_start_2_cpus):
-    setting = RayExecutor.create_settings(timeout_s=30,
-                                          placement_group_timeout_s=5)
-    hjob = RayExecutor(setting, num_hosts=1, num_slots=4)
+def test_cross_rank():
+    settings = MiniSettings()
+    coord = Coordinator(settings)
+    assert coord.world_size == 0
+    assert coord.hoststring == ""
+    ranks = list(range(12))
+
+    for r in ranks:
+        if r < 5:
+            coord.register("host1", world_rank=r)
+        elif r < 9:
+            coord.register("host2", world_rank=r)
+        else:
+            coord.register("host3", world_rank=r)
+
+    rank_to_info = coord.finalize_registration()
+    assert len(rank_to_info) == len(ranks)
+    # check that there is only 1 rank with cross_size == 1, cross_rank == 0
+    cross_size_1 = list(info for info in rank_to_info.values()
+                        if info["HOROVOD_CROSS_SIZE"] == 1)
+    assert len(cross_size_1) == 1
+    assert cross_size_1[0]["HOROVOD_CROSS_RANK"] == 0
+    # check that there is only 2 rank with cross_size == 2
+    cross_size_2 = list(info for info in rank_to_info.values()
+                        if info["HOROVOD_CROSS_SIZE"] == 2)
+    assert len(cross_size_2) == 2
+
+    # check that if cross_size == 2, set(cross_rank) == 0,1
+    assert set(d["HOROVOD_CROSS_RANK"] for d in cross_size_2) == {0, 1}
+
+    # check that there is 9 rank with cross_size = 3
+    cross_size_3 = list(info for info in rank_to_info.values()
+                        if info["HOROVOD_CROSS_SIZE"] == 3)
+    assert len(cross_size_3) == 9
+
+
+# Used for Pytest parametrization.
+parameter_str = "num_workers,num_hosts,num_workers_per_host"
+ray_executor_parametrized = [(4, None, None), (None, 1, 4)]
+
+
+@pytest.mark.parametrize(parameter_str, ray_executor_parametrized)
+def test_infeasible_placement(ray_start_2_cpus, num_workers, num_hosts,
+                              num_workers_per_host):
+    setting = RayExecutor.create_settings(
+        timeout_s=30, placement_group_timeout_s=5)
+    hjob = RayExecutor(
+        setting,
+        num_workers=num_workers,
+        num_hosts=num_hosts,
+        num_workers_per_host=num_workers_per_host)
     with pytest.raises(TimeoutError):
         hjob.start()
     hjob.shutdown()
 
 
-@pytest.mark.skipif(torch.cuda.device_count() < 4,
-                    reason="GPU test requires 4 GPUs")
-@pytest.mark.skipif(not torch.cuda.is_available(),
-                    reason="GPU test requires CUDA.")
+@pytest.mark.skipif(
+    torch.cuda.device_count() < 4, reason="GPU test requires 4 GPUs")
+@pytest.mark.skipif(
+    not torch.cuda.is_available(), reason="GPU test requires CUDA.")
 def test_gpu_ids(ray_start_4_cpus_4_gpus):
     original_resources = ray.available_resources()
     setting = RayExecutor.create_settings(timeout_s=30)
-    hjob = RayExecutor(setting, num_hosts=1, num_slots=4, use_gpu=True)
+    hjob = RayExecutor(
+        setting, num_hosts=1, num_workers_per_host=4, use_gpu=True)
     hjob.start()
     worker_handles = hjob.workers
     all_envs = ray.get([h.env_vars.remote() for h in worker_handles])
@@ -125,10 +172,15 @@ def test_horovod_mixin(ray_start_2_cpus):
     assert ray.get(actor.env_vars.remote())["TEST"] == str(DUMMY_VALUE)
 
 
-def test_local(ray_start_4_cpus):
+@pytest.mark.parametrize(parameter_str, ray_executor_parametrized)
+def test_local(ray_start_4_cpus, num_workers, num_hosts, num_workers_per_host):
     original_resources = ray.available_resources()
     setting = RayExecutor.create_settings(timeout_s=30)
-    hjob = RayExecutor(setting, num_hosts=1, num_slots=4)
+    hjob = RayExecutor(
+        setting,
+        num_workers=num_workers,
+        num_hosts=num_hosts,
+        num_workers_per_host=num_workers_per_host)
     hjob.start()
     hostnames = hjob.execute(lambda _: socket.gethostname())
     assert len(set(hostnames)) == 1, hostnames
@@ -138,7 +190,9 @@ def test_local(ray_start_4_cpus):
 
 @pytest.mark.skipif(
     not gloo_built(), reason='Gloo is required for Ray integration')
-def test_ray_init(ray_start_4_cpus):
+@pytest.mark.parametrize(parameter_str, ray_executor_parametrized)
+def test_ray_init(ray_start_4_cpus, num_workers, num_hosts,
+                  num_workers_per_host):
     original_resources = ray.available_resources()
 
     def simple_fn(worker):
@@ -148,7 +202,11 @@ def test_ray_init(ray_start_4_cpus):
 
     setting = RayExecutor.create_settings(timeout_s=30)
     hjob = RayExecutor(
-        setting, num_hosts=1, num_slots=4, use_gpu=torch.cuda.is_available())
+        setting,
+        num_workers=num_workers,
+        num_hosts=num_hosts,
+        num_workers_per_host=num_workers_per_host,
+        use_gpu=torch.cuda.is_available())
     hjob.start()
     result = hjob.execute(simple_fn)
     assert len(set(result)) == 4
@@ -158,7 +216,9 @@ def test_ray_init(ray_start_4_cpus):
 
 @pytest.mark.skipif(
     not gloo_built(), reason='Gloo is required for Ray integration')
-def test_ray_exec_func(ray_start_4_cpus):
+@pytest.mark.parametrize(parameter_str, ray_executor_parametrized)
+def test_ray_exec_func(ray_start_4_cpus, num_workers, num_hosts,
+                       num_workers_per_host):
     def simple_fn(num_epochs):
         import horovod.torch as hvd
         hvd.init()
@@ -166,7 +226,11 @@ def test_ray_exec_func(ray_start_4_cpus):
 
     setting = RayExecutor.create_settings(timeout_s=30)
     hjob = RayExecutor(
-        setting, num_hosts=1, num_slots=4, use_gpu=torch.cuda.is_available())
+        setting,
+        num_workers=num_workers,
+        num_hosts=num_hosts,
+        num_workers_per_host=num_workers_per_host,
+        use_gpu=torch.cuda.is_available())
     hjob.start()
     result = hjob.run(simple_fn, args=[0])
     assert len(set(result)) == 1
@@ -175,7 +239,9 @@ def test_ray_exec_func(ray_start_4_cpus):
 
 @pytest.mark.skipif(
     not gloo_built(), reason='Gloo is required for Ray integration')
-def test_ray_exec_remote_func(ray_start_4_cpus):
+@pytest.mark.parametrize(parameter_str, ray_executor_parametrized)
+def test_ray_exec_remote_func(ray_start_4_cpus, num_workers, num_hosts,
+                              num_workers_per_host):
     def simple_fn(num_epochs):
         import horovod.torch as hvd
         hvd.init()
@@ -183,7 +249,11 @@ def test_ray_exec_remote_func(ray_start_4_cpus):
 
     setting = RayExecutor.create_settings(timeout_s=30)
     hjob = RayExecutor(
-        setting, num_hosts=1, num_slots=4, use_gpu=torch.cuda.is_available())
+        setting,
+        num_workers=num_workers,
+        num_hosts=num_hosts,
+        num_workers_per_host=num_workers_per_host,
+        use_gpu=torch.cuda.is_available())
     hjob.start()
     object_refs = hjob.run_remote(simple_fn, args=[0])
     result = ray.get(object_refs)
@@ -193,7 +263,9 @@ def test_ray_exec_remote_func(ray_start_4_cpus):
 
 @pytest.mark.skipif(
     not gloo_built(), reason='Gloo is required for Ray integration')
-def test_ray_executable(ray_start_4_cpus):
+@pytest.mark.parametrize(parameter_str, ray_executor_parametrized)
+def test_ray_executable(ray_start_4_cpus, num_workers, num_hosts,
+                        num_workers_per_host):
     class Executable:
         def __init__(self, epochs):
             import horovod.torch as hvd
@@ -206,19 +278,47 @@ def test_ray_executable(ray_start_4_cpus):
 
     setting = RayExecutor.create_settings(timeout_s=30)
     hjob = RayExecutor(
-        setting, num_hosts=1, num_slots=4, use_gpu=torch.cuda.is_available())
+        setting,
+        num_workers=num_workers,
+        num_hosts=num_hosts,
+        num_workers_per_host=num_workers_per_host,
+        use_gpu=torch.cuda.is_available())
     hjob.start(executable_cls=Executable, executable_args=[2])
     result = hjob.execute(lambda w: w.rank_epoch())
     assert set(result) == {0, 2, 4, 6}
     hjob.shutdown()
 
 
+@pytest.mark.skipif(
+    not gloo_built(), reason='Gloo is required for Ray integration')
+def test_ray_deprecation(ray_start_4_cpus):
+    class Executable:
+        def __init__(self, epochs):
+            import horovod.torch as hvd
+            self.hvd = hvd
+            self.epochs = epochs
+            self.hvd.init()
+
+        def rank_epoch(self):
+            return self.hvd.rank() * self.epochs
+
+    setting = RayExecutor.create_settings(timeout_s=30)
+    hjob = RayExecutor(
+        setting,
+        num_hosts=1,
+        num_slots=2,
+        cpus_per_slot=2,
+        use_gpu=torch.cuda.is_available())
+    hjob.start(executable_cls=Executable, executable_args=[2])
+    result = hjob.execute(lambda w: w.rank_epoch())
+    assert set(result) == {0, 2}
+    hjob.shutdown()
+
+
 def _train(batch_size=32, batch_per_iter=10):
-    import torch.backends.cudnn as cudnn
     import torch.nn.functional as F
     import torch.optim as optim
     import torch.utils.data.distributed
-    from torchvision import models
     import horovod.torch as hvd
     import timeit
 
@@ -246,20 +346,26 @@ def _train(batch_size=32, batch_per_iter=10):
         loss.backward()
         optimizer.step()
 
-    time = timeit.timeit(benchmark_step, number=batch_per_iter)
+    timeit.timeit(benchmark_step, number=batch_per_iter)
     return hvd.local_rank()
 
 
 @pytest.mark.skipif(
     not gloo_built(), reason='Gloo is required for Ray integration')
-def test_horovod_train(ray_start_4_cpus):
+@pytest.mark.parametrize(parameter_str, ray_executor_parametrized)
+def test_horovod_train(ray_start_4_cpus, num_workers, num_hosts,
+                       num_workers_per_host):
     def simple_fn(worker):
         local_rank = _train()
         return local_rank
 
     setting = RayExecutor.create_settings(timeout_s=30)
     hjob = RayExecutor(
-        setting, num_hosts=1, num_slots=4, use_gpu=torch.cuda.is_available())
+        setting,
+        num_workers=num_workers,
+        num_hosts=num_hosts,
+        num_workers_per_host=num_workers_per_host,
+        use_gpu=torch.cuda.is_available())
     hjob.start()
     result = hjob.execute(simple_fn)
     assert set(result) == {0, 1, 2, 3}
@@ -269,4 +375,5 @@ def test_horovod_train(ray_start_4_cpus):
 if __name__ == "__main__":
     import pytest
     import sys
+
     sys.exit(pytest.main(["-v", __file__] + sys.argv[1:]))